Reorganize directories to be more nix-appropriate
Organizing modules and services in a way that makes sense instead of a
way just to make stow make sense. There's really no reason that the
modules and services need stowed in the first place since the home
manager imports are relative to the real file.

Adds a bunch of config for i3, vim, polybar, alacritty.